backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
bcrypt==4.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
mysql-connector-python==8.0.33
pymysql==1.0.3
cryptography
gevent>=1.4
